fix(metrics): use Keychain credentials for /v1/models so context max is correct
The "ctx" indicator on the session list was permanently stuck at 100%
for sessions running Opus on macOS. The context value itself was right
(~310K tokens) but contextMaxTokens fell back to 200,000 instead of
the model's actual 1,000,000 limit, so contextPercent was clamped at
100 (155% before clamping).

Root cause: anthropic-models.ts only read the OAuth token from
~/.claude/.credentials.json. Newer Claude Code on macOS stores it in
the Keychain (which we already handle for /v1/usage in
anthropic-usage.ts but had not extended to /v1/models). With no token,
the /v1/models fetch returned an empty map and getMaxInputTokens()
fell back to the 200K default.

Fix:
- Extract the file → Keychain credential lookup into a shared utility
  utils/claude-credentials.ts and use it from both anthropic-models.ts
  and anthropic-usage.ts.
- Update User-Agent in anthropic-models.ts from
  "claude-code/2.0.32" to "cchub/<version>" (matches the change made
  to anthropic-usage.ts in v0.1.93).

Verified on a session showing ~321K context tokens against the
claude-opus-4-7 model:
  before:  contextMaxTokens=200000  contextPercent=100
  after:   contextMaxTokens=1000000 contextPercent=32.2

Linux is unaffected: getClaudeAccessToken still reads
~/.claude/.credentials.json first, and the Keychain branch is gated
on process.platform === 'darwin'.
